+++ views/main/index.ejs
@@ -51,12 +51,14 @@
             <div class="row" float="right">
             <% if(isAuthenticated){ %>
                 <span><%= currentUser.username %> 환영합니다.</span>
-                <button type="button" onclick="location.href='/logout'">Logout</button>
+                <button type="button"><a href="/users/logout">Logout</a></button>
             <% } else { %>
                 <div class="login">
-                    <form action="/login" method="post">
-                      ID <input type="text" name="username" id="username" value="<%= username %>" placeholder="id">
-                      PASSWORD <input type="password" name="password" id="password" placeholder="password">
+                    <form action="/users/login" method="post">
+                       <label for="username">ID</label>
+                       <input type="text" name="username" id="username" value="<%= username %>" placeholder="id">
+                       <label for="password">PASSWORD</label>
+                       <input type="password" name="password" id="password" placeholder="password">
                       Login<input class="" type="submit" value="Submit">
                     </form>
                   </div>
