+++ views/users/signup.ejs
@@ -88,7 +88,7 @@
               <label for="name">이름</label>
               <input type="text" name='name' id='name' value="<%= user.name %>">
               <% if(errors.name){ %>
-                <span><%= erros.name.message %></span>
+                <span><%= errors.name.message %></span>
               <% } %>
             </div>
             <div>
